Move the amf and swf controllers into the api namespace
