+++ .rubocop_todo.yml
@@ -28,7 +28,7 @@ Lint/AssignmentInCondition:
 # Offense count: 4
 Lint/HandleExceptions:
   Exclude:
-    - 'app/controllers/amf_controller.rb'
+    - 'app/controllers/api/amf_controller.rb'
     - 'app/controllers/users_controller.rb'
 
 # Offense count: 692
